# backend/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  handler.InventoryListResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/model.InventoryItem'
        type: array
      limit:
        type: integer
      page:
        type: integer
      total:
        type: integer
    type: object
...
  model.BillingUnit:
    enum:
    - per_day
    - per_night
    type: string
    x-enum-varnames:
    - BillingUnitPerDay
    - BillingUnitPerNight
  model.BodySize:
    enum:
    - small
    - medium
    - large
    type: string
    x-enum-varnames:
    - BodySizeSmall
    - BodySizeMedium
    - BodySizeLarge
...
  model.HospitalizationPlan:
    properties:
      billing_unit:
        $ref: '#/definitions/model.BillingUnit'
      body_size:
        $ref: '#/definitions/model.BodySize'
      clinic_id:
        type: integer
      created_at:
        type: string
      description:
        type: string
      id:
        type: integer
      is_active:
        type: boolean
      name:
        type: string
      price:
        type: number
      sort_order:
        type: integer
      updated_at:
        type: string
    type: object
...
  model.Insurance:
    properties:
      clinic_id:
        type: integer
      contact_phone:
        type: string
      coverage_rate:
        type: integer
      created_at:
        type: string
      description:
        type: string
      id:
        type: integer
      is_active:
        type: boolean
      name:
        type: string
      sort_order:
        type: integer
      updated_at:
        type: string
    type: object
  model.InventoryCategory:
    enum:
    - medicine
    - consumable
    - food
    - other
    type: string
    x-enum-varnames:
    - InventoryCategoryMedicine
    - InventoryCategoryConsumable
    - InventoryCategoryFood
    - InventoryCategoryOther
  model.InventoryItem:
    properties:
      category:
        $ref: '#/definitions/model.InventoryCategory'
      clinic_id:
        type: integer
      created_at:
        type: string
      expiry_date:
        type: string
      id:
        type: integer
      last_restocked:
        type: string
      location:
        type: string
      min_stock_level:
        type: integer
      name:
        type: string
      quantity:
        type: integer
      status:
        $ref: '#/definitions/model.InventoryStatus'
      supplier:
        type: string
      unit:
        type: string
      updated_at:
        type: string
    type: object
  model.InventoryStatus:
    enum:
    - sufficient
    - low
    - out_of_stock
    type: string
    x-enum-varnames:
    - InventoryStatusSufficient
    - InventoryStatusLow
    - InventoryStatusOutOfStock
...
host: localhost:8080
info:
  contact: {}
  description: 動物病院 電子カルテシステム API（45テーブル・専用マスタテーブル版）
  title: Animal Ekarte API
  version: "2.0"
paths:
...
  /inventory:
    get:
      consumes:
      - application/json
      description: 在庫アイテムの一覧をページネーション付きで取得する
      parameters:
      - description: 'ページ番号 (default: 1)'
        in: query
        name: page
        type: integer
      - description: '件数 (1-100, default: 20)'
        in: query
        name: limit
        type: integer
      - description: カテゴリフィルター
        in: query
        name: category
        type: string
      - description: ステータスフィルター (sufficient, low, out_of_stock)
        in: query
        name: status
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.InventoryListResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: 在庫一覧取得
      tags:
      - Inventory
    post:
      consumes:
      - application/json
      description: 新しい在庫アイテムを作成する
      parameters:
      - description: 在庫情報
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/model.InventoryItem'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/model.InventoryItem'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: 在庫作成
      tags:
      - Inventory
  /inventory/{id}:
    get:
      consumes:
      - application/json
      description: 指定IDの在庫アイテムを取得する
      parameters:
      - description: 在庫ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.InventoryItem'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: 在庫詳細取得
      tags:
      - Inventory
    put:
      consumes:
      - application/json
      description: 指定IDの在庫アイテムを更新する
      parameters:
      - description: 在庫ID
        in: path
        name: id
        required: true
        type: integer
      - description: 更新する在庫情報
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/model.InventoryItem'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.InventoryItem'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: 在庫更新
      tags:
      - Inventory
...
  /masters/hospitalization-plans:
    get:
      consumes:
      - application/json
      description: 登録されている入院プランの一覧を返す
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.HospitalizationPlan'
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: 入院プラン一覧取得
      tags:
      - Masters
    post:
      consumes:
      - application/json
      description: 新しい入院プランを作成する
      parameters:
      - description: 入院プラン情報
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/model.HospitalizationPlan'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/model.HospitalizationPlan'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: 入院プラン作成
      tags:
      - Masters
  /masters/hospitalization-plans/{id}:
    delete:
      consumes:
      - application/json
      description: 指定IDの入院プランを削除する
      parameters:
      - description: 入院プランID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: 入院プラン削除
      tags:
      - Masters
    put:
      consumes:
      - application/json
      description: 指定IDの入院プランを更新する
      parameters:
      - description: 入院プランID
        in: path
        name: id
        required: true
        type: integer
      - description: 入院プラン情報
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/model.HospitalizationPlan'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.HospitalizationPlan'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: 入院プラン更新
      tags:
      - Masters
  /masters/insurances:
    get:
      description: クリニックに登録されている保険の一覧を返す。
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.Insurance'
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: 保険一覧取得
      tags:
      - InsuranceMasters
    post:
      consumes:
      - application/json
      description: 新規保険を登録する。
      parameters:
      - description: 保険情報
        in: body
        name: insurance
        required: true
        schema:
          $ref: '#/definitions/model.Insurance'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/model.Insurance'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: 保険作成
      tags:
      - InsuranceMasters
  /masters/insurances/{id}:
    delete:
      description: 指定IDの保険を削除する。
      parameters:
      - description: 保険ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: 保険削除
      tags:
      - InsuranceMasters
    put:
      consumes:
      - application/json
      description: 指定IDの保険情報を更新する。
      parameters:
      - description: 保険ID
        in: path
        name: id
        required: true
        type: integer
      - description: 更新する保険情報
        in: body
        name: insurance
        required: true
        schema:
          $ref: '#/definitions/model.Insurance'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Insurance'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: 保険更新
      tags:
      - InsuranceMasters
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"